requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Jsonpify==1.5.0
PyYAML==5.3.1
SQLAlchemy==1.3.22
PyMySQL==1.0.2
PyJWT==2.0.0


